Add cli-utils file
